Migrated classes related to Intel GPU compatibility to utils.Intel_GPU Python module.
